size_t fixed

+++ 0x01-python-if_else_loops_functions/linked_lists.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include "lists.h"
@@ -7,7 +8,8 @@
  * @h: pointer to head of list
  * Return: number of nodes
  */
-unsigned int print_listint(const listint_t *h)
+
+size_t print_listint(const listint_t *h)
 {
     const listint_t *current;
     unsigned int n; /* number of nodes */
